Fixed party command bugs

+++ src/com/raeic/embarker/party/commands/DismissCommand.java
@@ -17,12 +17,12 @@ public class DismissCommand implements CommandExecutor {
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         if (args.length > 2 || args.length < 1 || (args.length == 2 && !args[1].equalsIgnoreCase("confirm"))) {
             sender.sendMessage("Usage: /dismiss <player> (confirm)");
-            return false;
+            return true;
         }
 
         if (!(sender instanceof Player)) {
             sender.sendMessage("This command can only be ran by a player.");
-            return false;
+            return true;
         }
 
         Player p = (Player) sender;
@@ -56,21 +56,32 @@ public class DismissCommand implements CommandExecutor {
         }
 
         if (p.getName().equalsIgnoreCase(playerToDismiss.getName())) {
-            p.sendMessage("You can't dismiss yourself. You can leave with /leave " + p.getName() + ".");
+            p.sendMessage("You can't dismiss yourself. You can leave with /leave.");
             return true;
         }
 
         if (args.length == 2 && args[1].equalsIgnoreCase("confirm")) {
+            // Notify the party that the player was dismissed
             for (String oldPartyPlayerUniqueId : party.getPartyPlayersUniqueId()) {
-                Player oldPartyPlayer = Bukkit.getPlayer(oldPartyPlayerUniqueId);
+                Player oldPartyPlayer = Bukkit.getPlayer(UUID.fromString(oldPartyPlayerUniqueId));
+
                 if (oldPartyPlayer != null && oldPartyPlayer.isOnline()) {
-                    oldPartyPlayer.sendMessage(playerToDismiss.getName() + " was dismissed from " + p.getName() + "'s party.");
+                    if (oldPartyPlayerUniqueId.equals(userToDismissUniqueId)) {
+                        oldPartyPlayer.sendMessage("You were dismissed from " + p.getName() + "'s party.");
+                    } else {
+                        oldPartyPlayer.sendMessage(playerToDismiss.getName() + " was dismissed from " + p.getName() + "'s party.");
+                    }
                 }
             }
 
-            PartyPlayer dismissedPartyPlayer = PartyPlayer.findOne(userToDismissUniqueId);
-            if (dismissedPartyPlayer != null) {
-                dismissedPartyPlayer.delete();
+            // Disband the party of there are only two players. Otherwise just remove the dismissed player.
+            if (party.getPartyPlayersUniqueId().size() <= 2) {
+                party.disband();
+            } else {
+                PartyPlayer dismissedPartyPlayer = PartyPlayer.findOne(userToDismissUniqueId);
+                if (dismissedPartyPlayer != null) {
+                    dismissedPartyPlayer.delete();
+                }
             }
 
         } else {
